revert

+++ Plugins/BattleInfoPlugin/Models/Raw/battle_midnight_sp_midnight.cs
@@ -8,8 +8,8 @@
         public int api_deck_id { get; set; }
         public int[] api_ship_ke { get; set; }
         public int[] api_ship_lv { get; set; }
-		public decimal[] api_nowhps { get; set; }
-		public decimal[] api_maxhps { get; set; }
+		public int[] api_nowhps { get; set; }
+		public int[] api_maxhps { get; set; }
         public int[][] api_eSlot { get; set; }
         public int[][] api_eKyouka { get; set; }
         public int[][] api_fParam { get; set; }
