+++ Plugins/BattleInfoPlugin/Models/Raw/combined_battle_midnight_battle.cs
@@ -8,10 +8,10 @@
         public int api_deck_id { get; set; }
         public int[] api_ship_ke { get; set; }
         public int[] api_ship_lv { get; set; }
-		public decimal[] api_nowhps { get; set; }
-		public decimal[] api_maxhps { get; set; }
-		public decimal[] api_nowhps_combined { get; set; }
-		public decimal[] api_maxhps_combined { get; set; }
+		public int[] api_nowhps { get; set; }
+		public int[] api_maxhps { get; set; }
+		public int[] api_nowhps_combined { get; set; }
+		public int[] api_maxhps_combined { get; set; }
         public int[][] api_eSlot { get; set; }
         public int[][] api_eKyouka { get; set; }
         public int[][] api_fParam { get; set; }
