+++ Plugins/BattleInfoPlugin/Models/Raw/practice_midnight_battle.cs
@@ -6,10 +6,10 @@
     public class practice_midnight_battle
     {
         public int api_deck_id { get; set; }
-		public decimal[] api_nowhps { get; set; }
+		public int[] api_nowhps { get; set; }
         public int[] api_ship_ke { get; set; }
         public int[] api_ship_lv { get; set; }
-		public decimal[] api_maxhps { get; set; }
+		public int[] api_maxhps { get; set; }
         public int[][] api_eSlot { get; set; }
         public int[][] api_eKyouka { get; set; }
         public int[][] api_fParam { get; set; }
